+++ src/main/kotlin/di/readerModule.kt
@@ -2,11 +2,13 @@ package di
 
 import org.koin.dsl.module
 import presentation.ui_io.FloatReader
+import presentation.ui_io.InputReader
 import presentation.ui_io.IntReader
 import presentation.ui_io.StringReader
 
 val readerModule = module{
-    single { IntReader() }
+    single<InputReader<Int>> { IntReader() }
     single { StringReader() }
-    single { FloatReader() }
+    single <InputReader<Float>>{ FloatReader()  }
+
 }
