+++ src/main/kotlin/presentation/GetMealByIdUI.kt
@@ -1,11 +1,11 @@
 package presentation
 
 import logic.usecase.GetMealByIdUseCase
-import presentation.ui_io.IntReader
+import presentation.ui_io.InputReader
 
 class GetMealByIdUI(
     private val getMealByIdUseCase: GetMealByIdUseCase,
-    private val intReader: IntReader
+    private val intReader: InputReader<Int>
 ) : ChangeFoodMoodLauncher {
     override fun launchUI() {
         print("enter id of the meal : ")
