+++ src/main/kotlin/presentation/GetMealByNameUI.kt
@@ -1,11 +1,11 @@
 package presentation
 
 import logic.usecase.GetMealBySearchForNameUseCase
-import presentation.ui_io.StringReader
+import presentation.ui_io.InputReader
 
 class GetMealByNameUI(
     private val getMealBySearchForNameUseCase: GetMealBySearchForNameUseCase,
-    private val stringReader: StringReader
+    private val stringReader: InputReader<String>
 ) : ChangeFoodMoodLauncher {
 
     override fun launchUI() {
